+++ packages/Button/test/Button.spec.js
@@ -2,6 +2,7 @@
 
 import React from "react";
 import { render, fireEvent, act } from "@testing-library/react";
+import { axe } from "jest-axe";
 import Button from "../src";
 
 const noop = () => {};
@@ -123,5 +124,10 @@ describe("Button", () => {
     act(() => buttonRef.focus());
 
     expect(document.activeElement.tagName.toLowerCase()).toEqual("button");
+  });
+
+  it("should not fail any accessibility tests", async () => {
+    const { container } = renderComponent();
+    expect(await axe(container)).toHaveNoViolations();
   });
 });