+++ packages/Uploader/tests/Uploader.spec.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
+import { axe } from "jest-axe";
 import Uploader from "../src";
 
 function renderComponent() {
@@ -18,5 +19,10 @@ describe("Uploader", () => {
     expect(getByText("Drop files to upload here or")).toBeInTheDocument();
     expect(getByText("choose from your computer")).toBeInTheDocument();
     expect(getByTestId("uploader-dropZone-uploadIcon")).toBeInTheDocument();
+  });
+
+  it("should not fail any accessibility tests", async () => {
+    const { container } = renderComponent();
+    expect(await axe(container)).toHaveNoViolations();
   });
 });